Added a side drawer for smaller screens, fixed position of the web3Modal.
