+++ src/App.js
@@ -1,16 +1,16 @@
 import React, { useState, useEffect, useContext } from "react";
 import HarvestContext from "./Context/HarvestContext";
-import styled, { ThemeProvider } from "styled-components";
+import { ThemeProvider } from "styled-components";
 import { Row, Col } from "styled-bootstrap-grid";
-import { createGlobalStyle } from "styled-components";
-import { reset } from "styled-reset";
 import harvest from "./lib/index.js";
 import Loadable from "react-loadable";
-import { darkTheme, lightTheme, fonts } from "./styles/appStyles.js";
+import { darkTheme, lightTheme} from "./styles/appStyles.js";
 import axios from "axios";
 
 // images
 import logo from "./assets/gif_tractor.gif";
+//styles
+import {Topbar, GlobalStyle, Brand, Panel, Container} from './styles/AppJsStyles';
 
 // components
 import TabContainer from "./components/tabContainer/TabContainer";
@@ -21,223 +21,12 @@ import WelcomeText from "./components/WelcomeText";
 import CheckBalance from "./components/checkBalance/CheckBalance";
 import TokenMessage from "./components/statusMessages/TokenMessage";
 import HarvestAndStakeMessage from "./components/statusMessages/HarvestAndStakeMessage";
+import Sidedrawer from './components/userSettings/sidedrawer/Sidedrawer';
 
 const { ethers } = harvest;
-const GlobalStyle = createGlobalStyle`
-  ${reset}
- 
-  
-  html {
-    /* 1rem = 10px */
-    font-size: 62.5%;
-    height: 100%;
-  }
 
 
 
-  body {
-    height: 100%;
-    background-color: ${(props) => props.theme.style.bodyBackground};
-  }
-
-
-  code {
-    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-      monospace;
-  }
-
-  
-
-
-  input[type="number"] {
-    -moz-appearance: textfield;
-    background-color: ${(props) => props.theme.style.lightBackground};
-    border: 0.2rem solid #363636;
-    font-size: 1.4rem;
-    color: ${(props) => props.theme.style.primaryFontColor};
-    width: 6rem;
-    text-align: center;
-    border-radius: 0.5rem;
-    padding: 0.3rem 0.7rem;
-    @media(max-width: 1400px) {
-      width: 6rem;
-    }
-    @media(max-width: 1280px) {
-      width: 5rem;
-    }
-  }
-
-  input[type="number"]::-webkit-inner-spin-button,
-  input[type="number"]::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    appearance: none;
-  }
-
-  .button {
-    background: ${(props) => props.theme.style.highlight};
-    border: ${(props) => props.theme.style.smallBorder};
-    box-shadow: ${(props) => props.theme.style.buttonBoxShadow};
-    box-sizing: border-box;
-    border-radius: 0.8rem;
-    padding: 0.5rem 1rem;
-    cursor: pointer;
-    font-family: TechnaSans;
-    color: ${(props) => props.theme.style.buttonFontColor};
-    font-size: 1.2rem;
-
-    &.ghost {
-      background: transparent;
-      border: 0px;
-      box-shadow: none;
-      color: ${(props) => props.theme.style.linkColor};
-      font-family: ${fonts.headerFont};
-      padding: 0px;
-    }
-
-    &.alert {
-      background-color: ${(props) => props.theme.style.alertColor}
-    }
-  }
-
-  .spread-row {
-    justify-content: space-between;
-  }
-
-  div[data-name="row"] {
-    margin-bottom: 1.5rem;
-  }
-`;
-
-// App
-const Brand = styled.div`
-  padding-right: 1rem;
-  padding-top: 2rem;
-  display: flex;
-  align-items: flex-start;
-  margin-bottom: 4rem;
-  height: 2.5rem;
-
-  img {
-    width: 3rem;
-    height: 3rem;
-    margin-right: 1rem;
-    margin-left: 0.5rem;
-  }
-
-  span {
-    color: ${(props) => props.theme.style.brandTextColor};
-    font-family: ${fonts.contentFont};
-    font-size: 2.5rem;
-  }
-
-  @media (min-width: 1500px) {
-    margin: 3rem 0;
-  }
-`;
-
-const Topbar = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  .fa-user-cog {
-    color: ${(props) => props.theme.style.brandTextColor};
-    font-size: 2.5rem;
-    padding-top: 2rem;
-    padding-right: 1rem;
-    position: relative;
-
-    &:hover {
-      cursor: pointer;
-      top: 0.2rem;
-    }
-  }
-`;
-
-const Panel = styled.div`
-  position: relative;
-  padding: 2.5rem 2.5rem;
-  border: ${(props) => props.theme.style.mainBorder};
-  border-radius: 1rem;
-  border-top-left-radius: 0rem;
-  margin-top: -1.5rem;
-  background-color: ${(props) => props.theme.style.panelBackground};
-  z-index: 100;
-  box-shadow: ${(props) => props.theme.style.panelBoxShadow};
-  
-
-  &.four-corner {
-    border-top-left-radius: 1rem;
-    background-color: #1d1d1d;
-    color: ${(props) => props.theme.style.primaryFontColor};
-    font-size: 1.6rem;
-    font-family: TechnaSans;
-  }
-
-  //Radio Modal
-  .flexible-modal {
-    position: absolute;
-    z-index: 51;
-    background-color: #ddd;
-    height: 2rem;
-    border: 1px solid black;
-    border-radius .5rem;
-    background-color: ${(props) => props.theme.style.highlight};
-    box-shadow: ${(props) => props.theme.style.panelTabBoxShadow};
-    margin-bottom: 0;
-  }
-  
-  .flexible-modal-mask {
-    display: none;
-    
-  }
-  .flexible-modal-resizer {
-    height: 2rem;
-  }
-  
-  
-  .flexible-modal-drag-area{
-    background-color: transparent;
-    position:absolute;
-    cursor:grab;
-    height: 2rem;
-    right:0;
-    top:0;
-  }
-  .flexible-modal-drag-area-bottom{
-    height: 3rem;
-    cursor:grab;
-    position:absolute;
-    bottom: 0;
-  }
-  .flexible-modal-drag-area-right{
-    position:absolute;
-    height: 5rem;
-    width: 2rem;
-    top: 0;
-    right: 0;
-    cursor:grab;
-  }
-  .flexible-modal-drag-area-left{
-    position:absolute;
-    height: 5rem;
-    width: 2rem;
-    top: 0;
-    left: 0;
-    cursor:grab;
-  }
-
-
- 
-`;
-
-const Container = styled.div`
-  width: 100%;
-  max-width: 1200px;
-  margin: 0 auto;
-  @media (max-width: 1140px) {
-    width: 95%;
-  }
-`;
 
 const ErrorModal = Loadable({
   loader: () => import("./components/ErrorModal"),
@@ -252,6 +41,7 @@ function App() {
   const [exchangeRates, setExchangeRates] = useState({});
   const [currentExchangeRate, setCurrentExchangeRate] = useState(1);
   //for currency conversion
+  const [openDrawer, setOpenDrawer] = useState(false);
   const [settingsOpen, setSettingsOpen] =useState(false);
   const [isConnecting, setIsConnecting] = useState(false);
   const [isRefreshing, setRefreshing] = useState(false);
@@ -373,6 +163,10 @@ function App() {
   const toggleUserSettings = () => {
     setSettingsOpen(!settingsOpen)
   }
+  const toggleSideDrawer = () => {
+    setOpenDrawer(!openDrawer)
+  }
+  
 
   const setConnection = (provider, signer, manager) => {
     setState({
@@ -478,11 +272,16 @@ function App() {
         prettyBalance,
         convertStandardNumber,
         settingsOpen,
-        toggleUserSettings
+        toggleUserSettings,
+        openDrawer,
+        toggleSideDrawer
+        
       }}
     >
       <ThemeProvider theme={state.theme === "dark" ? darkTheme : lightTheme}>
         <GlobalStyle />
+        {openDrawer ? <Sidedrawer /> : null}
+        
         <Container>
           <Row>
             <Col col>
@@ -493,6 +292,7 @@ function App() {
                 </Brand>
                 <i onClick={toggleUserSettings} className="fas fa-user-cog"></i>
                 {settingsOpen ? <SettingsModal /> : ''}
+                <i className="fas fa-bars" onClick={toggleSideDrawer}></i>
                 
               </Topbar>
             </Col>
