+++ src/components/tabContainer/TabContainerStyles.jsx
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
-import {fonts} from '../../styles/appStyles';
+import styled from "styled-components";
+import { fonts } from "../../styles/appStyles";
 
 export const PanelTab = styled.div`
   margin-right: 0.75rem;
@@ -10,12 +10,9 @@ export const PanelTab = styled.div`
   padding: 0.75rem 2rem 2.25rem 2rem;
   background-color: ${(props) => props.theme.style.highlight};
   box-shadow: ${(props) => props.theme.style.panelTabBoxShadow};
-  
+
   cursor: pointer;
   color: ${(props) => props.theme.style.buttonFontColor};
-
-  
-  
 
   a {
     color: ${(props) => props.theme.style.panelTabLinkColor};
@@ -23,45 +20,18 @@ export const PanelTab = styled.div`
     font-family: ${fonts.contentFont};
     font-size: 2.4rem;
     position: relative;
-    top: .1rem;
-    @media(max-width: 500px) {
+    top: 0.1rem;
+    @media (max-width: 500px) {
       font-size: 1.5rem;
-      top: .3rem;
-    }
-    
-   
-  }
-  @media(max-width: 700px) {
-    font-size: 1.6rem;
-    padding: 0.75rem 1rem 2.2rem 1rem;
-    position: relative;
-    top: .1rem;
-    
-  }
-  @media(max-width: 550px) {
-    margin-right: .5rem;
-  }
-  @media(max-width: 380px) {
-    font-size: 1.2rem;
-    padding: 0.75rem .5rem 2rem .5rem;
-    position: relative;
-    margin-right: .0rem;
-    top: .5rem;
-    a {
-      top: .4rem;
+      top: 0.3rem;
     }
   }
-  @media(max-width: 333px) {
-    
-  }
-  
 
   &.wiki-tab {
     position: relative;
     background-color: ${(props) => props.theme.style.wikiTabBackground};
     top: 0.4rem;
     margin-left: 1rem;
-    
 
     &:hover {
       top: 0rem;
@@ -71,33 +41,27 @@ export const PanelTab = styled.div`
       color: ${(props) => props.theme.style.primaryFontColor};
       font-size: 1.9rem;
       position: relative;
-      top: .1rem;
+      top: 0.1rem;
     }
-    @media(max-width: 575px) {
-      
-      margin-left: .5rem;
-     
-      
+    @media (max-width: 1160px) {
+      display: none;
     }
-   
-    @media(max-width: 500px) {
-      top: 1.3rem;
-      margin-left: .5rem;
-      a {
-        font-size: 1.5rem;
-        top: -.1rem;;
-      }
-      
+  }
+
+  
+
+  &.radio-tab {
+    position: relative;
+    font-size: 1.9rem;
+    top: 0.6rem;
+    margin-left: 1rem;
+    font-family: ${fonts.contentFont};
+    &:hover {
+      top: 0rem;
     }
-    @media(max-width: 380px) {
-      margin-left: 0;
-      a {
-        font-size: 1.4rem;
-      }
-      
-    };
-    @media(max-width: 333px) {
-      margin-right: .3rem;
+
+    @media (max-width: 1160px) {
+      display: none;
     }
   }
 
@@ -117,76 +81,12 @@ export const PanelTab = styled.div`
       font-family: ${fonts.contentFont};
       font-size: 1.9rem;
       position: relative;
-      top: -.1rem;
-      
+      top: -0.1rem;
     }
-    @media(max-width: 675px) {
-      
-      margin-left: .5rem;
-      
-     
-      
-    }
-   
-    @media(max-width: 500px) {
-      top: 1.3rem;
-      margin-left: .5rem;
-      p {
-        font-size: 1.5rem;
-        
-      }
-      
-    }
-    @media(max-width: 380px) {
-      margin-left: 0;
-       {
-        font-size: 1.4rem;
-      }
-      
-    };
-    @media(max-width: 333px) {
-      margin-right: .3rem;
+    @media (max-width: 1160px) {
+      display: none;
     }
   }
-   
-
-    &.radio-tab {
-      position: relative;
-      font-size: 1.9rem;
-      top: .6rem;
-      margin-left: 1rem;
-      font-family: ${fonts.contentFont};
-      &:hover {
-        top: 0rem;
-      }
-
-      @media(max-width: 700px) {
-        margin-left: .5rem;
-        p {
-          font-size: 1.9rem;
-        }
-        
-      }
-
-      @media(max-width: 500px) {
-        top: 1.4rem;
-        margin-left: 1rem;
-        p {
-          font-size: 1.5rem;
-        }
-        
-      }
-      @media(max-width: 380px) {
-        margin-left: 0;
-        p {
-          font-size: 1.4rem;
-        }
-        
-      }
-      @media(max-width: 333px) {
-        margin-left: .3rem;
-      }
-  
 `;
 
 export const PanelTabContainer = styled.div`
