+++ src/components/userSettings/sidedrawer/backdrop/BackdropStyles.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const BackdropContainer = styled.div`
+    position: fixed;
+    pointer-events: auto;
+    z-index: 550;
+    background-color: rgba(0,0,0,0.4);
+    height: 100%;
+    width: 105vw;
+
+`;
